requirements.txt
requests
prettytable
python-stdnum
pydantic
typer
pytest
